+++ packages/contracts/src/interactor-operation-catalog.test.ts
@@ -16,7 +16,7 @@ const device = {
 const available = { available: true } as const;
 const unavailable = { available: false, reason: 'unsupported-device-kind' } as const;
 
-test('binds every requested operation the facts admitted, driving the resolved interactor', async () => {
+test('binds every operation the facts admitted, driving the resolved interactor', async () => {
   const back = vi.fn(async () => undefined);
   const home = vi.fn(async () => undefined);
   const resolveInteractor = vi.fn(async () => ({ back, home }) as unknown as Interactor);
@@ -27,7 +27,6 @@ test('binds every requested operation the facts admitted, driving the resolved i
     signal,
     resolveInteractor,
     facts: { back: available, home: available },
-    operations: ['back', 'home'],
   });
 
   expect(operations.back).toBeTypeOf('function');
@@ -38,31 +37,32 @@ test('binds every requested operation the facts admitted, driving the resolved i
   expect(home).toHaveBeenCalledOnce();
 });
 
-test('skips a requested operation the facts refused', () => {
+test('skips an operation the facts refused', () => {
   const operations = bindAdmittedLocalInteractorOperations({
     device,
     signal: new AbortController().signal,
     resolveInteractor: vi.fn(),
     facts: { back: available, home: unavailable },
-    operations: ['back', 'home'],
   });
 
   expect(operations.back).toBeTypeOf('function');
   expect(operations.home).toBeUndefined();
 });
 
-test('never binds an admitted operation the caller did not request', () => {
+// #1955 review: the facts a caller passes are the ONLY source of truth for what binds — there is
+// no separate, caller-maintained operation list that could drift from them. An operation the
+// facts admit binds even if the caller only meant to name a couple of others.
+test('binds every admitted operation the facts name, not just the ones a caller had in mind', () => {
   const operations = bindAdmittedLocalInteractorOperations({
     device,
     signal: new AbortController().signal,
     resolveInteractor: vi.fn(),
     facts: { back: available, home: available, tvRemote: available },
-    operations: ['back'],
   });
 
   expect(operations.back).toBeTypeOf('function');
-  expect(operations.home).toBeUndefined();
-  expect(operations.tvRemote).toBeUndefined();
+  expect(operations.home).toBeTypeOf('function');
+  expect(operations.tvRemote).toBeTypeOf('function');
 });
 
 test('treats an operation missing from a partial facts map as unavailable', () => {
@@ -71,14 +71,13 @@ test('treats an operation missing from a partial facts map as unavailable', () =
     signal: new AbortController().signal,
     resolveInteractor: vi.fn(),
     facts: { back: available },
-    operations: ['back', 'tvRemote'],
   });
 
   expect(operations.back).toBeTypeOf('function');
   expect(operations.tvRemote).toBeUndefined();
 });
 
-test('a provider binding drives its own resolved interactor for every requested, admitted operation', async () => {
+test('a provider binding drives its own resolved interactor for every admitted operation', async () => {
   const tvRemote = vi.fn(async () => undefined);
   const resolveInteractor = vi.fn(() => ({ tvRemote }) as unknown as Interactor);
 
@@ -87,7 +86,6 @@ test('a provider binding drives its own resolved interactor for every requested,
     signal: new AbortController().signal,
     resolveInteractor,
     facts: { tvRemote: available },
-    operations: ['back', 'home', 'tvRemote'],
   });
 
   expect(operations.back).toBeUndefined();
@@ -103,7 +101,6 @@ test('a provider binding fails closed when its exact owner exposes no interactor
     signal: new AbortController().signal,
     resolveInteractor: () => undefined,
     facts: { keyboardDismiss: available },
-    operations: ['keyboardDismiss'],
   });
 
   await expect(operations.keyboardDismiss?.({})).rejects.toMatchObject({
