+++ src/daemon/handlers/session.ts
@@ -1,8 +1,8 @@
 import { dispatchCommand } from '../../core/dispatch.ts';
 import { PUBLIC_COMMANDS } from '../../command-catalog.ts';
 import type { AndroidAdbExecutor } from '../../platforms/android/adb-executor.ts';
-import { publicPlatformString, type DeviceInfo } from '@agent-device/kernel/device';
-import type { DaemonInvokeFn, DaemonRequest, DaemonResponse, SessionState } from '../types.ts';
+import { publicPlatformString } from '@agent-device/kernel/device';
+import type { DaemonInvokeFn, DaemonRequest, DaemonResponse } from '../types.ts';
 import { SessionStore } from '../session-store.ts';
 import { contextFromFlags } from '../context.ts';
 import { handleReleaseMaterializedPathsCommand } from './session-app-source-deployment.ts';
@@ -10,14 +10,13 @@ import { requireSessionOrExplicitSelector, resolveCommandDevice } from './sessio
 import { errorResponse, requireCommandSupported } from './response.ts';
 import { recordSessionAction } from './handler-utils.ts';
 import { handleRuntimeCommand } from './session-runtime-command.ts';
-import { resolveBoundKeyboardRuntime } from '../keyboard-runtime.ts';
 import { requireRuntimeBinding, requireRuntimeFacts } from './session-runtime-admission.ts';
 import { handleOpenCommand } from './session-open.ts';
 import { composeOpenWithInitialSnapshot } from './session-open-foreground.ts';
 import {
-  resolveAndroidPackageForOpen,
-  resolveSessionAppBundleIdForTarget,
-} from '../../platform-runtime-open-target.ts';
+  handleKeyboardCommand,
+  handleTriggerAppEventCommand,
+} from './session-selector-dispatch.ts';
 import { handleCloseCommand } from './session-close.ts';
 import { handleSessionAppDeploymentCommand } from './session-app-deployment-route.ts';
 import { runBatchCommands } from './session-batch.ts';
@@ -28,9 +27,7 @@ import { handleSessionReplayCommands } from './session-replay.ts';
 import { handleSessionScriptPublication } from './session-script-publication.ts';
 import { handleDoctorCommand } from './session-doctor.ts';
 import { handlePrepareCommand } from './session-prepare.ts';
-import { resolveRefFrameEffect } from '../daemon-command-registry.ts';
 import type { DescriptorSessionRouteCommandName } from '../../core/command-descriptor/registry.ts';
-import { expireRefFrame } from '../ref-frame.ts';
 import { LeaseRegistry } from '../lease-registry.ts';
 import type { LeaseLifecycleProvider } from '@agent-device/contracts/device';
 import type {
@@ -42,108 +39,6 @@ import type { DeviceClaimReconciler } from '../device-claims.ts';
 import type { AppLogAdmissionLedger } from '../app-log-admission-ledger.ts';
 import type { ScreenRecordingAdmissionLedger } from '../screen-recording-admission-ledger.ts';
 import type { PlatformRequestScope } from '@agent-device/contracts/platform';
-
-/** What `runSessionOrSelectorDispatch`'s `execute` thunk reports: either the early-exit response
- * an admission refusal produces, or the result to record and return. */
-type SessionCommandExecutionOutcome =
-  | Readonly<{ ok: false; response: DaemonResponse }>
-  | Readonly<{ ok: true; result: Record<string, unknown> | void }>;
-
-/**
- * The one orchestration every session/selector-route leaf shares: guard, resolve the device,
- * admit-then-execute via the caller's own strategy, expire the ref frame if the command mutates,
- * derive and record the next session. `execute` is where the two strategies session-route
- * commands use today diverge — {@link legacySessionDispatchExecute} for a still-legacy command
- * (capability gate, then `dispatchCommand`), or a bind-and-execute thunk like `keyboard`'s below
- * for a migrated one — everything around it is identical either way, so it lives here once
- * instead of once per command.
- */
-// fallow-ignore-next-line complexity
-async function runSessionOrSelectorDispatch(params: {
-  req: DaemonRequest;
-  sessionName: string;
-  sessionStore: SessionStore;
-  command: string;
-  positionals: string[];
-  recordPositionals?: string[];
-  deriveNextSession?: (
-    session: SessionState,
-    result: Record<string, unknown> | void,
-    device: DeviceInfo,
-  ) => Promise<SessionState> | SessionState;
-  execute: (
-    device: DeviceInfo,
-    session: SessionState | undefined,
-  ) => Promise<SessionCommandExecutionOutcome>;
-}): Promise<DaemonResponse> {
-  const {
-    req,
-    sessionName,
-    sessionStore,
-    command,
-    positionals,
-    recordPositionals,
-    deriveNextSession,
-    execute,
-  } = params;
-  const session = sessionStore.get(sessionName);
-  const flags = req.flags ?? {};
-  const guard = requireSessionOrExplicitSelector(command, session, flags);
-  if (guard) return guard;
-
-  const device = await resolveCommandDevice({
-    session,
-    flags,
-    ensureReady: true,
-  });
-  const outcome = await execute(device, session);
-  if (!outcome.ok) return outcome.response;
-
-  // ADR 0014 side-effect seam for session/selector-route leaves (keyboard
-  // dismiss/enter/return, push, trigger-app-event). Expire the frame before
-  // recording when the classification says this request mutates; keyboard
-  // status/get resolve to `preserve` and leave the frame untouched.
-  if (session && resolveRefFrameEffect(req) === 'may-invalidate') {
-    expireRefFrame(session);
-  }
-
-  const result = outcome.result;
-  if (session) {
-    const nextSession = deriveNextSession
-      ? await deriveNextSession(session, result, device)
-      : session;
-    recordSessionAction(sessionStore, nextSession, req, command, result ?? {}, {
-      positionals: recordPositionals ?? positionals,
-    });
-    if (nextSession !== session) {
-      sessionStore.set(sessionName, nextSession);
-    }
-  }
-  return { ok: true, data: result ?? {} };
-}
-
-/** The still-legacy `execute` thunk: the capability gate and `dispatchCommand` call
- * `runSessionOrSelectorDispatch` always ran inline before `execute` existed as a parameter. Every
- * remaining unmigrated session-route command passes this until its own migration replaces it with
- * a bind-and-execute thunk, same as `keyboard`'s {@link handleKeyboardCommand} below. */
-function legacySessionDispatchExecute(
-  command: string,
-  positionals: string[],
-  req: DaemonRequest,
-  logPath: string,
-): (
-  device: DeviceInfo,
-  session: SessionState | undefined,
-) => Promise<SessionCommandExecutionOutcome> {
-  return async (device, session) => {
-    const unsupported = requireCommandSupported(command, device);
-    if (unsupported) return { ok: false, response: unsupported };
-    const result = await dispatchCommand(device, command, positionals, req.flags?.out, {
-      ...contextFromFlags(logPath, req.flags, session?.appBundleId, session?.trace?.outPath),
-    });
-    return { ok: true, result };
-  };
-}
 
 // fallow-ignore-next-line complexity
 async function handleClipboardCommand(params: {
@@ -291,112 +186,6 @@ const handleSessionReplayCommandGroup: SessionCommandHandler = async ({
     retainDeviceExecutionLock,
     throwIfCanceled,
   });
-
-/**
- * A dismiss/enter/return sent with no session and no explicit iOS selector would target whatever
- * app happens to be foreground on a later-resolved device, silently. Refuse it up front rather
- * than let admission and binding run against a target the caller never named.
- */
-function requireForegroundIosKeyboardSession(
-  session: SessionState | undefined,
-  keyboardAction: string | undefined,
-  flags: DaemonRequest['flags'],
-): DaemonResponse | undefined {
-  const needsForegroundIosApp =
-    keyboardAction === 'dismiss' || keyboardAction === 'enter' || keyboardAction === 'return';
-  if (session || !needsForegroundIosApp || flags?.platform !== 'ios') return undefined;
-  return errorResponse(
-    'SESSION_NOT_FOUND',
-    'iOS keyboard action requires an active session so the target app stays foregrounded. Run open first.',
-  );
-}
-
-/**
- * `keyboard`'s migrated `execute` thunk (ADR 0019 §9): bind whichever action-selected use the
- * parsed action names, then run it. Replaces the resolve-then-record shape
- * `runSessionOrSelectorDispatch` now owns — this only supplies what the generic route's
- * `dispatchCommand` cannot: the bound runtime and the action-specific execution context.
- */
-function keyboardSessionExecute(
-  positionals: string[],
-  inspectFacts: SessionCommandParams['inspectFacts'],
-  bindDevice: SessionCommandParams['bindDevice'],
-  req: DaemonRequest,
-  logPath: string,
-): (
-  device: DeviceInfo,
-  session: SessionState | undefined,
-) => Promise<SessionCommandExecutionOutcome> {
-  return async (device, session) => {
-    const bound = await resolveBoundKeyboardRuntime({
-      device,
-      positionals,
-      inspectFacts,
-      bindDevice,
-    });
-    if (!bound.ok) return { ok: false, response: bound.response };
-    const dispatchContext = {
-      ...contextFromFlags(logPath, req.flags, session?.appBundleId, session?.trace?.outPath),
-      surface: session?.surface,
-    };
-    const result = await bound.execute(dispatchContext);
-    return { ok: true, result };
-  };
-}
-
-async function handleKeyboardCommand(params: SessionCommandParams): Promise<DaemonResponse> {
-  const { req, sessionName, logPath, sessionStore, inspectFacts, bindDevice } = params;
-  const positionals = req.positionals ?? [];
-
-  const foregroundGuard = requireForegroundIosKeyboardSession(
-    sessionStore.get(sessionName),
-    positionals[0]?.trim().toLowerCase(),
-    req.flags ?? {},
-  );
-  if (foregroundGuard) return foregroundGuard;
-
-  return await runSessionOrSelectorDispatch({
-    req,
-    sessionName,
-    sessionStore,
-    command: PUBLIC_COMMANDS.keyboard,
-    positionals,
-    execute: keyboardSessionExecute(positionals, inspectFacts, bindDevice, req, logPath),
-  });
-}
-
-async function handleTriggerAppEventCommand(params: SessionCommandParams): Promise<DaemonResponse> {
-  const { req, sessionName, logPath, sessionStore } = params;
-  const positionals = req.positionals ?? [];
-  return await runSessionOrSelectorDispatch({
-    req,
-    sessionName,
-    sessionStore,
-    command: PUBLIC_COMMANDS.triggerAppEvent,
-    positionals,
-    execute: legacySessionDispatchExecute(
-      PUBLIC_COMMANDS.triggerAppEvent,
-      positionals,
-      req,
-      logPath,
-    ),
-    deriveNextSession: async (session, result) => {
-      const eventUrl = typeof result?.eventUrl === 'string' ? result.eventUrl : undefined;
-      const nextAppBundleId = eventUrl
-        ? ((await resolveSessionAppBundleIdForTarget(
-            session.device,
-            eventUrl,
-            session.appBundleId,
-            resolveAndroidPackageForOpen,
-          )) ?? session.appBundleId)
-        : session.appBundleId;
-      return {
-        ...session,
-        appBundleId: nextAppBundleId,
-      };
-    },
-  });
-}
 
 /**
  * Descriptor-driven exhaustive dispatch table for the daemon's `session`
